# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Teste",
        "description": "Descrição detalhada da API de Teste",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Lista todas as atividades",
                "description": "Retorna uma lista de todas as atividades",
                "operationId": "53e7a0b67f6097ed90245000764b2e5e",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Activity"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Cria uma nova atividade",
                "description": "Cria uma nova atividade com os dados fornecidos",
                "operationId": "8aa12114961e22767ecdad7c24be367e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Activity"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitação inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/activities/{id}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Mostra uma atividade específica",
                "description": "Retorna os detalhes de uma atividade específica",
                "operationId": "6d75cdd5535b9b12358448ec198faba4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da atividade a ser visualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Atividade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Atualiza uma atividade existente",
                "description": "Atualiza uma atividade existente com os dados fornecidos",
                "operationId": "064379988a5fe2c974ae5ab18f0cddfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da atividade a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Activity"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Atividade não encontrada"
                    },
                    "400": {
                        "description": "Solicitação inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Exclui uma atividade existente",
                "description": "Exclui uma atividade existente dado um ID",
                "operationId": "d822403ad9a85feb951b1073858a072d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da atividade a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "404": {
                        "description": "Atividade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Activity": {
                "required": [
                    "user_id",
                    "activity_type_id",
                    "title",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID da atividade",
                        "type": "integer",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "ID do usuário relacionado à atividade",
                        "type": "integer"
                    },
                    "activity_type_id": {
                        "description": "ID do tipo de atividade",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Título da atividade",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição da atividade",
                        "type": "string",
                        "nullable": true
                    },
                    "start_date": {
                        "description": "Data de início da atividade",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "due_date": {
                        "description": "Data de vencimento da atividade",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "end_date": {
                        "description": "Data de término da atividade",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status da atividade",
                        "type": "string",
                        "enum": [
                            "open",
                            "completed"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}